maplocation/requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-filter==25.1
Pillow==11.2.1
psycopg2-binary==2.9.9  # PostgreSQL adapter
graphene-django==3.1.5  # GraphQL for Django